+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/ChatMember.kt
@@ -1,6 +1,8 @@
 package dev.inmo.tgbotapi.types.chat.member
 
 import dev.inmo.tgbotapi.abstracts.WithUser
+import dev.inmo.tgbotapi.types.chat.PreviewUser
+import dev.inmo.tgbotapi.types.chat.User
 import dev.inmo.tgbotapi.types.statusField
 import dev.inmo.tgbotapi.types.untilDateField
 import dev.inmo.tgbotapi.utils.RiskFeature
@@ -12,12 +14,12 @@ import kotlinx.serialization.builtins.serializer
 import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
-import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.jsonPrimitive
 
 @Serializable(ChatMemberSerializer::class)
 sealed interface ChatMember : WithUser {
+    override val user: PreviewUser
     @Serializable(StatusSerializer::class)
     enum class Status(
         val status: String,
@@ -28,7 +30,7 @@ sealed interface ChatMember : WithUser {
         Administrator("administrator", AdministratorChatMemberImpl.serializer()),
         Member("member", MemberChatMemberImpl.serializer(), { status, json -> status == "member" && json[untilDateField] ?.jsonPrimitive == null }),
         SubscriptionMember("member", SubscriptionMemberChatMemberImpl.serializer(), { status, json -> status == "member" && json[untilDateField] ?.jsonPrimitive != null }),
-        Restricted("restricted", RestrictedChatMember.serializer()),
+        Restricted("restricted", RestrictedMemberChatMember.serializer()),
         Left("left", LeftChatMemberImpl.serializer()),
         Kicked("kicked", KickedChatMember.serializer())
     }
@@ -74,7 +76,7 @@ object ChatMemberSerializer : KSerializer<ChatMember> {
             is AdministratorChatMemberImpl -> AdministratorChatMemberImpl.serializer().serialize(encoder, value)
             is SubscriptionMemberChatMemberImpl -> SubscriptionMemberChatMemberImpl.serializer().serialize(encoder, value)
             is MemberChatMemberImpl -> MemberChatMemberImpl.serializer().serialize(encoder, value)
-            is RestrictedChatMember -> RestrictedChatMember.serializer().serialize(encoder, value)
+            is RestrictedMemberChatMember -> RestrictedMemberChatMember.serializer().serialize(encoder, value)
             is LeftChatMemberImpl -> LeftChatMemberImpl.serializer().serialize(encoder, value)
             is KickedChatMember -> KickedChatMember.serializer().serialize(encoder, value)
         }