+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/ChatMemberUpdated.kt
@@ -5,6 +5,7 @@ import dev.inmo.tgbotapi.abstracts.WithUser
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.chat.Chat
 import dev.inmo.tgbotapi.types.chat.PreviewChat
+import dev.inmo.tgbotapi.types.chat.PreviewUser
 import dev.inmo.tgbotapi.types.chat.User
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -14,7 +15,7 @@ data class ChatMemberUpdated(
     @SerialName(chatField)
     override val chat: PreviewChat,
     @SerialName(fromField)
-    override val user: User,
+    override val user: PreviewUser,
     @SerialName(dateField)
     val date: TelegramDate,
     @SerialName(oldChatMemberField)
@@ -27,4 +28,7 @@ data class ChatMemberUpdated(
     val viaChatFolderInviteLink: Boolean? = false,
     @SerialName(viaJoinRequestField)
     val viaJoinRequest: Boolean? = false
-) : WithPreviewChat, WithUser
+) : WithPreviewChat, WithUser {
+    val member: PreviewUser
+        get() = newChatMemberState.user
+}