+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/KickedChatMember.kt
@@ -1,16 +1,17 @@
 package dev.inmo.tgbotapi.types.chat.member
 
 import dev.inmo.tgbotapi.types.*
+import dev.inmo.tgbotapi.types.chat.PreviewUser
 import dev.inmo.tgbotapi.types.chat.User
 import kotlinx.serialization.*
 
 @Serializable
 data class KickedChatMember(
     @SerialName(userField)
-    override val user: User,
+    override val user: PreviewUser,
     @SerialName(untilDateField)
     override val untilDate: TelegramDate? = null
-) : BannedChatMember {
+) : RestrictedChatMember {
     @SerialName(statusField)
     @Required
     @EncodeDefault