+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/member/SubscriptionMemberChatMemberImpl.kt
@@ -1,6 +1,7 @@
 package dev.inmo.tgbotapi.types.chat.member
 
 import dev.inmo.tgbotapi.types.*
+import dev.inmo.tgbotapi.types.chat.PreviewUser
 import dev.inmo.tgbotapi.types.chat.User
 import korlibs.time.DateTime
 import kotlinx.serialization.*
@@ -8,7 +9,7 @@ import kotlinx.serialization.*
 @Serializable
 data class SubscriptionMemberChatMemberImpl(
     @SerialName(userField)
-    override val user: User,
+    override val user: PreviewUser,
     @SerialName(untilDateField)
     @Serializable(TelegramDateSerializer::class)
     override val untilDate: TelegramDate